Add snippet for tac-ing all logs including gzipped
Use this with grep -m<count> to search for recent messages.  Use
sparingly: This has a potential to download the entire uncompressed log
over SSH to local machine.  (Using with grep -m ensures that the
pipeline will be terminated with SIGPIPE before it has a chance to
download a lot of data.)

Think 10x before using this with PSSH.